backend/requirements.txt
# fastapi==0.115.2
# uvicorn[standard]==0.30.6
# pydantic-settings==2.4.0
# python-dotenv==1.0.1
# loguru==0.7.2

# fastapi
# uvicorn[standard]

langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface

google-generativeai
huggingface-hub
sentence-transformers

chromadb

python-dotenv
pydantic
httpx
tiktoken

docx2txt
python-docx
